+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added tween layers for controlling multiple tweens at a time.
 - Added option to tween using unscaled delta time.
-- Added WaitForTween custom yield instruction.
+- Added WaitForTween and WaitForTweenLayer as custom yield instructions.
 - Added JetBrain annotations.
 - Replaced exceptions with Debug.Warning.